Only shuffle A and AAAA records on cache hits
We previously added logic to shuffle records when we got a cache hit.
This introduced a bug because we were shuffling CNAME records, whose
order should be maintained.

This change fixes that by only shuffling the sub section of records in
the response starting from the first A or AAAA record and going until
the end.